tst: move testing utilities
