+++ tests/test_regression.py
@@ -1,6 +1,4 @@
 """Testing conversion with conversion saved on datalad"""
-from tempfile import mkdtemp
-import os
 import os.path as op
 import json
 from glob import glob
@@ -13,36 +11,10 @@ try:
 except ImportError:
 	have_datalad = False
 
-from heudiconv.cli.run import main as runner
 import heudiconv
+from heudiconv.cli.run import main as runner
+from .utils import fetch_data, gen_heudiconv_args
 
-
-def gen_heudiconv_args(datadir, outdir, subject, heuristic_file, xargs=None):
-	heuristic = op.realpath(op.join(
-								op.dirname(heudiconv.__file__),
-					        	'..',
-			        			'heuristics',
-			        			heuristic_file))
-
-	args = ["-d", op.join(datadir, 'sourcedata/{subject}/*/*/*.tgz'),
-			"-c", "dcm2niix",
-			"-o", outdir,
-			"-s", subject,
-			"-f", heuristic,
-			"--bids",
-			]
-	if xargs:
-		args += xargs
-
-	return args
-
-def fetch_data(tmpdir, subject):
-	"""Fetches some test dicoms"""
-	targetdir = os.path.join(tmpdir, 'QA')
-	api.install(path=targetdir,
-				source='///dbic/QA')
-	api.get('{}/sourcedata/{}'.format(targetdir, subject))
-	return targetdir
 
 @pytest.mark.parametrize('subject', ['sub-sid000143'])
 @pytest.mark.parametrize('heuristic', ['dbic_bids.py'])
